==> notebooks/2_transformation.ipynb <==
import findspark
findspark.init()

# PySpark is the Spark API for Python. In this lab, we use PySpark to initialize the spark context. 
from pyspark import SparkContext, SparkConf
from pyspark.sql import SparkSession
from pyspark.sql import SQLContext
from pyspark.sql import functions as F
from pyspark.sql.types import StructType,StructField, StringType, IntegerType, FloatType, BinaryType, DateType


# Criar o contexto do spark
sc = SparkContext()

# Instancia o criador de sessao do spark
spark = (SparkSession.builder
                     .master("local[7]")
                     .appName("Aceleração PySpark - Capgemini"))

schema_airports = StructType([
    StructField("faa",  StringType(),  True),
    StructField("name", StringType(),  True),
    StructField("lat",  FloatType(),   True),
    StructField("lon",  FloatType(),   True),
    StructField("alt",  IntegerType(), True),
    StructField("tz",   FloatType(), True),
    StructField("dst",  StringType(),  True),
    StructField("qa_faa",  StringType(),  True),
    StructField("qa_name",  StringType(),  True),
    StructField("qa_lat",  StringType(),  True),
    StructField("qa_lon",  StringType(),  True),
    StructField("qa_alt",  StringType(),  True),
    StructField("qa_tz",  StringType(),  True),
    StructField("qa_dst",  StringType(),  True)
])

schema_planes = StructType([
    StructField("tailnum",      StringType(),  True),
    StructField("year",         IntegerType(), True),
    StructField("type",         StringType(),  True),
    StructField("manufacturer", StringType(),  True),
    StructField("model",        StringType(),  True),
    StructField("engines",      IntegerType(), True),
    StructField("seats",        IntegerType(), True),
    StructField("speed",        IntegerType(), True),
    StructField("engine",       StringType(),  True),
    StructField("qa_tailnum",  StringType(),  True),
    StructField("qa_year",  StringType(),  True),
    StructField("qa_type",  StringType(),  True),
    StructField("qa_manufacturer",  StringType(),  True),
    StructField("qa_model",  StringType(),  True),
    StructField("qa_engines",  StringType(),  True),
    StructField("qa_seats",  StringType(),  True),
    StructField("qa_speed",  StringType(),  True),
    StructField("qa_enginge",  StringType(),  True)  
])

schema_flights = StructType([
    StructField("year",      IntegerType(), True),
    StructField("month",     IntegerType(), True),
    StructField("day",       IntegerType(), True),
    StructField("dep_time",  StringType(),  True),
    StructField("dep_delay", IntegerType(), True),
    StructField("arr_time",  StringType(),  True),
    StructField("arr_delay", IntegerType(), True),
    StructField("carrier",   StringType(),  True),
    StructField("tailnum",   StringType(),  True),
    StructField("flight",    StringType(),  True),
    StructField("origin",    StringType(),  True),
    StructField("dest",      StringType(),  True),
    StructField("air_time",  IntegerType(), True),
    StructField("distance",  IntegerType(), True),
    StructField("hour",      IntegerType(), True),
    StructField("minute",    IntegerType(), True),
    StructField("qa_year_month_day",    StringType(), True),
    StructField("qa_hour_minute",    StringType(), True),
    StructField("qa_dep_arr_time",    StringType(), True),
    StructField("qa_dep_arr_delay",    StringType(), True),
    StructField("qa_carrier",    StringType(), True),
    StructField("qa_tailnum",    StringType(), True),
    StructField("qa_flight",    StringType(), True),
    StructField("qa_origin_dest",    StringType(), True),
    StructField("qa_air_time",    StringType(), True),
    StructField("qa_distance",    StringType(), True),
    StructField("qa_distance_airtime",    StringType(), True)
])

airports_path_data = '../data/qa_airports.parquet'
planes_path_data   = '../data/qa_planes.parquet'
flights_path_data  = '../data/qa_flights.parquet'


airports = (spark.getOrCreate().read
                  .format("parquet")
                  .option("header", "true")
#                  .schema(schema_airports)
                  .load(airports_path_data))

planes   = (spark.getOrCreate().read
                  .format("parquet")
                  .option("header", "true")
                  .load(planes_path_data))

flights  = (spark.getOrCreate().read
                  .format("parquet")
                  .option("header", "true")
#                  .schema(schema_flights)
                  .load(flights_path_data))

airports.createOrReplaceTempView('airports')
planes.createOrReplaceTempView('planes')
flights.createOrReplaceTempView('flights')

planes = planes.withColumn('year',(
                    planes.year.cast(IntegerType())
)).withColumn('engines',(
                    planes.engines.cast(IntegerType())
)).withColumn('seats',(
                    planes.seats.cast(IntegerType())
)).withColumn('speed',(
                    planes.speed.cast(IntegerType())
))

planes.show(1)

def higher_than(col, condition):
    return (F.col(col) > condition )

def between(col, lower_condition, higher_condition):
    return ( F.col(col).between(lower_condition, higher_condition) )

def lower_than(col, condition):
    return (F.col(col) < condition )

def equal(col, condition):
    return (F.col(col) == condition )

## Airports Dataset 

airports_path_data = '../data/airports.csv'


#Schema
airports_schema = (StructType([ 
    StructField("faa",StringType(),True), 
    StructField("name",StringType(),True), 
    StructField("lat",FloatType(),True), 
    StructField("lon", FloatType(), True), 
    StructField("alt", IntegerType(), True), 
    StructField("tz", FloatType(), True),
    StructField("dst", StringType(), True )
                    ])
         )

#DataFrame
airports = spark.getOrCreate().read.csv(airports_path_data, header = True, schema = airports_schema)


airports.show(5)

airports.printSchema()

### Pergunta 1

airports = airports.withColumn('alt', (
                 F.when(lower_than('alt', 0), 0 )
                 .otherwise(F.col('alt'))
))

airports.where(airports.alt < 0).show()

### Pergunta 2

airports = airports.withColumn('dst', (
                    F.when( (airports.tz >= -7) &
                            (airports.tz <= -5)
                            ,'A')
                    .otherwise(F.col('dst'))
))

#airports.where((airports.tz == -5)).show(5)

### Pergunta 3

airports = airports.withColumn('dst',(
                    F.when(equal('dst', 'U'), 'A')
                    .otherwise(F.col('dst'))
))

airports.where(airports.dst == 'U').show()

### Pergunta 4

airports = airports.withColumn('region', (
                    F.when(lower_than('lon', -124), 'ALASKA')
                     .when(higher_than('lon', -50) | lower_than('lat',  24), 'OFFSHORE')
                     .when(lower_than('lon', -95) | equal('lon', -95), 'MAINLAND-WEST')
                     .when(higher_than('lon', -95), 'MAINLAND-EAST')
                     .otherwise(None)
))

airports.where((airports.lon.between(-124, -50)) &
               (airports.region != 'MAINLAND-WEST') &
               (airports.region != 'MAINLAND-EAST')).show()

### Pergunta 5

airports = airports.withColumn('type',(
                    F.when(F.col('name').rlike('Airport|Tradeport|Heliport|Airpor|Arpt' ), 'AP')
                     .when(F.col('name').rlike('Aerodrome'), 'AD')
                     .when(F.col('name').rlike('Airpark|Aero Park' ), 'AK')
                     .when(F.col('name').rlike('Station|Air Station'), 'AS')
                     .when(F.col('name').rlike('Field|Fld'), 'FL')
                     .otherwise(None)
))

airports.where((F.col('type') != 'AP') &
              (F.col('type') != 'AD') &
              (F.col('type') != 'AK') &
              (F.col('type') != 'AS') &
              (F.col('type') != 'FL') 
             ).show()

### Pergunta 6

#string_military = "Base", "Aaf", "AFs", "Ahp", "Afb", "LRRS", "Lrrs", "Arb", "Naf", "NAS", "Nas", "Jrb", "Ns", "As", "Cgas", "Angb"
string_military2 = ["Base", "Aaf", "AFs", "Ahp", "Afb", "LRRS", "Lrrs", "Arb", "Naf", "NAS", "Nas", "Jrb", "Ns",
"As", "Cgas", "Angb"]

#regex_string_military = ''
#regex_string_military = r'|'.join(map(lambda x : f".*(^| ){x}($| ).*", string_military))

regex_string_military2 = ''
regex_string_military2 = r'|'.join(map(lambda x : f".*(^| ){x}($| ).*", string_military2))


airports = airports.withColumn('military',(
                    F.when(F.col('name').rlike(regex_string_military2), True)
                    .otherwise(False)
))

#airports.filter(airports.military == False).show(5, truncate =False)

print(airports.where(F.col('military') == True).count())
print(airports.where(F.col('military') == False).count())

### Pergunta 7

airports = airports.withColumn('administration',(
                    F.when(F.col('name').rlike('International|Intl|Intercontinental'), 'I')
                    .when(F.col('name').rlike('National|Natl'), 'N')
                    .when(F.col('name').rlike('Regional|Reigonal|Rgnl|County|Metro|Metropolitan'), 'R')
                    .when(F.col('name').rlike('Municipal|Muni|City'), 'I')
))

airports.where(airports.administration.isNull()).show(10)

airports.write.mode('overwrite').parquet('../data/datasets_transformados/airports.parquet')

## Planes Dataset

path_data_planes = '../data/planes.csv'

#Schema
schema = (StructType([ 
    StructField("tailnum",StringType(),True), 
    StructField("year",IntegerType(),True), 
    StructField("type",StringType(),True), 
    StructField("manufacturer", StringType(), True), 
    StructField("model", StringType(), True), 
    StructField("engines", IntegerType(), True),
    StructField("seats", IntegerType(), True ),
    StructField("speed", IntegerType(), True),
    StructField("enginge", StringType(), True)
                    ])
         )

#DataFrame
planes = spark.getOrCreate().read.csv(path_data_planes, header = True, schema = schema)

### Pergunta 1

planes = planes.withColumn('tailchar',(
                 F.regexp_replace(F.substring(F.col('tailnum'), 2 , 6), '[0-9]', '')
))

### Pergunta 2

planes = planes.withColumn('year',(
                 F.when(F.col('year') == 0, 1996 )
                 .otherwise(F.col('year'))
))

### Pergunta 3

planes.sort('manufacturer', 'model', 'year')
print((planes.count(), len(planes.columns)))


#planes = planes.withColumn('manufacturer_model_year',(
#                 F.concat_ws('-', planes.manufacturer, planes.model, planes.year)
#))

aux_planes = planes.select('manufacturer', 'model', 'year').where(planes.year.isNotNull())
aux_planes = (aux_planes.withColumnRenamed('manufacturer', 'aux_manufacturer')
                        .withColumnRenamed('model', 'aux_model')
                        .withColumnRenamed('year', 'aux_year'))

aux_planes2 = planes.select('manufacturer', 'year').where(planes.year.isNotNull())
aux_planes2 = (aux_planes2.withColumnRenamed('manufacturer', 'aux_manufacturer')
                        .withColumnRenamed('year', 'aux_year'))

aux_planes = (aux_planes.groupBy('aux_manufacturer', 'aux_model')
                        .agg({'aux_year':'min',})
                        .withColumnRenamed('min(aux_year)', 'aux_year'))

aux_planes2 = (aux_planes2.groupBy('aux_manufacturer')
                          .agg({'aux_year':'min',})
                          .withColumnRenamed('min(aux_year)', 'aux_year'))

print((joined_planes.count(), len(joined_planes.columns)))

joined_planes = planes.join(aux_planes, 
            (planes.manufacturer == aux_planes.aux_manufacturer) &
            (planes.model == aux_planes.aux_model),
            'left'
)

joined_planes = joined_planes.withColumn('year',(
                        F.when((joined_planes.year.isNull()) &
                               (joined_planes.aux_year.isNotNull()), joined_planes.aux_year
                              )
                         .otherwise(joined_planes.year)
))

joined_planes = joined_planes.drop('aux_manufacturer', 'aux_model', 'aux_year')



joined_planes = joined_planes.join(aux_planes2, 
            (joined_planes.manufacturer == aux_planes2.aux_manufacturer),
            'left'
)

joined_planes = joined_planes.withColumn('year',(
                        F.when((joined_planes.year.isNull()) &
                               (joined_planes.aux_year.isNotNull()), joined_planes.aux_year
                              )
                         .otherwise(joined_planes.year)
))

planes = joined_planes

### Pergunta 4

planes = planes.withColumn('age',(
                 F.when(planes.year.isNotNull(), 2022 - planes.year )
                 .otherwise(None)
                 ))

### Pergunta 5

planes = planes.withColumn('type',(
                 F.when(planes.type == 'Fixed wing multi engine', 'MULTI_ENG')
                 .when(planes.type == 'Fixed Wing single engine', 'SINGLE_ENG')
                 .when(planes.type == 'Rotorcraft', 'ROTORCRAFT')
))

### Pergunta 6

planes.select('manufacturer').distinct().sort('manufacturer').show(truncate = False)

planes = planes.withColumn('manufacturer',(
                 F.when(planes.manufacturer.startswith('AIRBUS'), 'AIRBUS')
                  .when(planes.manufacturer.startswith('BOMBARDIER'), 'BOMBARDIER')
                  .when(planes.manufacturer.startswith('MCDONNELL'), 'MCDONNELL DOUGLAS')
                  .when(planes.manufacturer.startswith('CIRRUS'), 'CIRRUS')
                  .when(planes.manufacturer.startswith('BARKER'), 'BARKER JACK')
                  .when(planes.manufacturer.startswith('ROBINSON'), 'ROBINSON HELICOPTER')
                  .when(planes.manufacturer.startswith('GULFSTREAM'), 'GULFSTREAM')
                  .otherwise(planes.manufacturer)
))

### Pergunta 7

planes.select('model').distinct().sort('model').show(truncate = False)

planes = planes.withColumn('model',(
                    F.when( planes.model.contains("("), F.regexp_replace(planes.model, '\([^()]*\)', '' ))
                    .otherwise(planes.model)
))
planes = planes.withColumn('model',(
            F.trim(planes.model)
))

planes.select('model').distinct().sort('model').show(5,truncate = False)

### Pergunta 8

planes.show(1)

planes = planes.withColumn('speed', (
                            F.when(planes.speed.isNull(), F.ceil((planes.seats)/(0.36)))
                            .otherwise(planes.speed)
))
planes.where(planes.speed.isNull()).show()

##verificar se está no intervalo possível do dataframe

### Pergunta 9

planes.select('enginge').distinct().show()
planes = planes.withColumn('engine_split',(
                 F.split(planes.enginge, '-')[1]
))

planes.select('engine_split').distinct().show()

planes = planes.withColumn('engine_type',(
                    F.when(planes.engine_split == 'jet', 'JET')
                    .when(planes.engine_split  == 'fan', 'FAN')
                    .when(planes.engine_split  == 'prop', 'PROP')
                    .when(planes.engine_split  == 'shaft', 'SHAFT')
                    .when(planes.enginge.endswith('Cycle'), 'CYCLE')
                    .otherwise(planes.enginge)
))

planes = planes.drop('engine_split')
planes.select('enginge', 'engine_type').distinct().show()

planes.write.mode('overwrite').parquet('../data/datasets_transformados/planes.parquet')

## Flights Dataset

flights_schema = (StructType([ 
    StructField("year",IntegerType(),True), 
    StructField("month",IntegerType(),True), 
    StructField("day",IntegerType(),True), 
    StructField("dep_time", StringType(), True), 
    StructField("dep_delay", IntegerType(), True), 
    StructField("arr_time", StringType(), True),
    StructField("arr_delay", IntegerType(), True ),
    StructField("carrier", StringType(), True),
    StructField("tailnum", StringType(), True),
    StructField("flight", StringType(), True),
    StructField("origin", StringType(), True),
    StructField("dest", StringType(), True),
    StructField("air_time", IntegerType(), True),
    StructField("distance", IntegerType(), True),
    StructField("hour", IntegerType(), True),
    StructField("minute", IntegerType(), True)
                    ])
                 )

flights = spark.getOrCreate().read.csv('../data/flights.csv', schema = flights_schema, header = True)

flights.show(5)
print(flights.count())

### Pergunta 1

flights = flights.withColumn('hour',(
                    F.when(flights.hour.isNull(), 0)
                    .otherwise(flights.hour)
))

flights = flights.withColumn('minute',(
                    F.when(flights.minute.isNull(), 0)
                    .otherwise(flights.minute)
))

### Pergunta 2

flights = flights.withColumn('hour',(
                    F.when(flights.hour == 24, 0)
                    .otherwise(flights.hour)
))

### Pergunta 3

flights = flights.withColumn('dep_datetime',(
                    F.date_format(F.concat_ws(' ', F.concat_ws('-', flights.year, flights.month, flights.day),
                                                    F.concat_ws(':', flights.hour, flights.minute)       
                                              ), ('yyyy-MM-dd hh:mm:00')
                                 )
))

flights = flights.withColumn('dep_datetime',(
                    F.to_timestamp( flights.dep_datetime, 'yyyy-MM-dd hh:mm:ss')
))

flights = flights.withColumn('dep_datetime', 
                    F.to_timestamp( F.concat_ws(' ', F.concat_ws('-', flights.year, flights.month, flights.day),
                                                     F.concat_ws(':', flights.hour, flights.minute, F.lit(00))       
                                               )
                                  )
)


flights.select('dep_datetime').show(5)

flights = flights.withColumn('month',(
                    F.when(F.length(flights.month) == 1, F.concat_ws('', F.lit('0'), flights.month))
                    .otherwise(flights.month)
))

flights = flights.withColumn('day',(
                    F.when(F.length(flights.day) == 1, F.concat_ws('', F.lit('0'), flights.day))
                    .otherwise(flights.day)
))

flights = flights.withColumn('hour',(
                    F.when(F.length(flights.hour) == 1, F.concat_ws('', F.lit('0'), flights.hour))
                    .otherwise(flights.hour)
))

flights = flights.withColumn('minute',(
                    F.when(F.length(flights.minute) == 1, F.concat_ws('', F.lit('0'), flights.minute))
                    .otherwise(flights.minute)
))

flights = flights.withColumn('aux_date',(
                   F.concat_ws('-', flights.year, flights.month, flights.day) 
))

flights = flights.withColumn('aux_hour',(
                    F.concat_ws(':', flights.hour, flights.minute, F.lit('00'))
))

flights = flights.withColumn('dep_datetime',(
                    F.concat_ws(' ', flights.aux_date, flights.aux_hour)
))

flights = flights.drop('aux_date', 'aux_hour')
#flights.select('aux_date', 'aux_hour', 'dep_datetime').show(5)

flights.select('dep_datetime').show(5)

### Pergunta 4

flights.show(2)

#flights.where(flights.dep_time == 'NA').show(5)

flights = flights.withColumn('dep_time',(
                    F.when(flights.dep_time == 'NA', F.concat_ws('',F.substring(flights.dep_datetime, 12,2), 
                                                                    F.substring(flights.dep_datetime, 14,2)))
                      .otherwise(flights.dep_time)
))

flights.select('dep_time').show(5)

### Pergunta 5 e 6

flights = flights.withColumn('dep_delay',(
                    F.when(flights.dep_delay.isNull(), 0)
                    .otherwise(flights.dep_delay)
))

flights = flights.withColumn('arr_delay',(
                    F.when(flights.arr_delay.isNull(), 0)
                    .otherwise(flights.arr_delay)
))

### Pergunta 7

flights = flights.drop('year', 'month', 'day', 'hour', 'minute')

### Pergunta 8

flights = flights.withColumn('air_time_projected',(
                    F.round((flights.distance * 0.1) + 20, 2)
))

flights = flights.withColumn('air_time_projected',(
                    flights.air_time_projected.cast(IntegerType())
))


### Pergunta 9
#### Montar parecido com a ideia do exercicio 3 planes 

flights = flights.withColumn('aux_origin_dest',(
                    F.concat_ws('-', flights.origin, flights.dest)
))

groupby_aux_origin_dest = (flights.groupBy('aux_origin_dest')
                                  .agg({'air_time': 'avg'})
                                  .withColumnRenamed('avg(air_time)', 'air_time_expected')
                                  .withColumnRenamed('aux_origin_dest', 'aux_origin_dest1')
                          )

flights = flights.join(groupby_aux_origin_dest,
             flights.aux_origin_dest == groupby_aux_origin_dest.aux_origin_dest1,
             'left'
            )

flights = flights.drop('aux_origin_dest1')
flights = flights.withColumn('air_time_expected',(
                            F.ceil(flights.air_time_expected)
))

flights.sort(flights.aux_origin_dest).show(3)

### Pergunta 10

flights = flights.withColumn('air_time',(
                        F.when(flights.air_time.isNull(), F.greatest('air_time_projected', 'air_time_expected'))
                         .otherwise(flights.air_time)
))


flights = flights.withColumn('air_time',( 
                  flights.air_time.cast(IntegerType())          
))


### Pergunta 11

flights = flights.withColumn('arr_time',(
                    F.when(flights.arr_time.isNull(), flights.dep_time + flights.air_time)
                    .otherwise(flights.arr_time)
))

flights.where(flights.arr_time> 60).show(5)

### Pergunta 12

flights = flights.withColumn('haul_duration',(
                    F.when((flights.air_time >= 20) &
                           (flights.air_time < 180),
                           'SHORT-HAUL'
                          )
                     .when((flights.air_time >= 180) &
                           (flights.air_time < 360),
                           'MEDIUM-HAUL'
                          )
                     .when((flights.air_time >= 360), 
                           'LONG-HAUL'
                          )
))

### Pergunta 13

flights.select('dep_datetime').show(2)

flights = flights.withColumn('dep_season',
F.when(F.col('dep_datetime').between(F.concat(F.year('dep_datetime'),F.lit('-03-20 15:33:00')),F.concat(F.year('dep_datetime'),F.lit('-06-21 10:14:00'))), 'SPRING')
 .when(F.col('dep_datetime').between(F.concat(F.year('dep_datetime'),F.lit('-06-21 10:14:00')),F.concat(F.year('dep_datetime'),F.lit('-09-23 02:04:00'))), 'SUMMER')
 .when(F.col('dep_datetime').between(F.concat(F.year('dep_datetime'),F.lit('-09-23 02:04:00')),F.concat(F.year('dep_datetime'),F.lit('-12-21 21:48:00'))), 'FALL')
 .otherwise('WINTER')
                                  )


flights.select('dep_season').groupBy('dep_season').count().show()


flights = flights.withColumn('dep_delay_category',(
                    F.when(flights.dep_delay < 0, 'ANTECIPATED')
                     .when(flights.dep_delay == 0, 'INTIME')
                     .when((flights.dep_delay > 0) & (flights.dep_delay < 60), 'MINOR')
                     .when(flights.dep_delay >= 60, 'MAJOR')
))

flights.write.mode('overwrite').parquet('../data/datasets_transformados/flights.parquet')
