==> notebooks/3_report.ipynb <==
import findspark
findspark.init()

# PySpark is the Spark API for Python. In this lab, we use PySpark to initialize the spark context. 
from pyspark import SparkContext, SparkConf
from pyspark.sql import SparkSession
from pyspark.sql import SQLContext
from pyspark.sql import functions as F
from pyspark.sql.types import StructType,StructField, StringType, IntegerType, FloatType, BinaryType, DateType


# Criar o contexto do spark
sc = SparkContext()

# Instancia o criador de sessao do spark
spark = (SparkSession.builder
                     .master("local[7]")
                     .appName("Aceleração PySpark - Capgemini"))

airports_path_data = '../data/qa_airports.parquet'
planes_path_data   = '../data/qa_planes.parquet'
flights_path_data  = '../data/qa_flights.parquet'


airports = (spark.getOrCreate().read
                  .format("parquet")
                  .option("header", "true")
                  .load(airports_path_data))

planes   = (spark.getOrCreate().read
                  .format("parquet")
                  .option("header", "true")
                  .load(planes_path_data))

flights  = (spark.getOrCreate().read
                  .format("parquet")
                  .option("header", "true")
                  .load(flights_path_data))

## Perguntas para Qualidade

### Unificando os datasets

planes  = planes.withColumnRenamed('tailnum', 'planes_tailnum').withColumnRenamed('qa_tailnum', 'planes_qa_tailnum')
flights = flights.withColumnRenamed('tailnum', 'flights_tailnum').withColumnRenamed('qa_tailnum', 'flights_qa_tailnum')

origin_to_join = airports.select('faa',
                                   'qa_faa',
                                   'qa_name',
                                   'qa_lat',
                                   'qa_lon',
                                   'qa_alt',
                                   'qa_tz',
                                   'qa_dst')

origin_to_join = (airports.withColumnRenamed('faa', 'origin_faa')
                         .withColumnRenamed('qa_faa', 'origin_qa_faa')
                         .withColumnRenamed('qa_name', 'origin_qa_name')
                         .withColumnRenamed('qa_lat', 'origin_qa_lat')
                         .withColumnRenamed('qa_lon', 'origin_qa_lon')
                         .withColumnRenamed('qa_alt', 'origin_qa_alt')
                         .withColumnRenamed('qa_tz', 'origin_qa_tz')
                         .withColumnRenamed('qa_dst', 'origin_qa_dst'))


dest_to_join = (airports.withColumnRenamed('faa', 'dest_faa')
                         .withColumnRenamed('qa_faa', 'dest_qa_faa')
                         .withColumnRenamed('qa_name', 'dest_qa_name')
                         .withColumnRenamed('qa_lat', 'dest_qa_lat')
                         .withColumnRenamed('qa_lon', 'dest_qa_lon')
                         .withColumnRenamed('qa_alt', 'dest_qa_alt')
                         .withColumnRenamed('qa_tz', 'dest_qa_tz')
                         .withColumnRenamed('qa_dst', 'dest_qa_dst'))

planes_to_join   = planes.select( 'planes_tailnum',
                                    'planes_qa_tailnum',
                                    'qa_year',
                                    'qa_type',
                                    'qa_manufacturer',
                                    'qa_model',
                                    'qa_engines',
                                    'qa_seats',
                                    'qa_speed',
                                    'qa_enginge')

flights_to_join  = flights.select( 'flights_tailnum',
                                    'origin',
                                    'dest',
                                    'qa_year_month_day',
                                    'qa_hour_minute',
                                    'qa_dep_arr_time',
                                    'qa_dep_arr_delay',
                                    'qa_carrier',
                                    'flights_qa_tailnum',
                                    'qa_flight',
                                    'qa_origin_dest',
                                    'qa_air_time',
                                    'qa_distance',
                                    'qa_distance_airtime')

df = flights_to_join.join(origin_to_join, 
                     (flights_to_join.origin == origin_to_join.origin_faa),                    
                    'left'
                    )

df = df.join(dest_to_join, 
                     (flights_to_join.dest == dest_to_join.dest_faa),                    
                    'left'
                    )

df = df.join(planes_to_join, 
                     (df.flights_tailnum == planes_to_join.planes_tailnum),
                    'left'
                    )

#### M - dado faltante
#### F  - indica que não respeita o formato esperado 
#### I   - indica que o valor excede o intervalo esperado
#### S - Indica que não tem exatamente 5 caracteres.
#### T

#### Formatando os dados para conter apenas valores M, F, I, S, T e null

for column in df.schema.names:
    if 'qa_' in column:
        df = df.withColumn(column, (
                          F.when(F.col(column).startswith('M'), 'M')
                           .when(F.col(column).startswith('F'), 'F')
                           .when(F.col(column).startswith('I'), 'I')
                           .when(F.col(column).startswith('S'), 'S')
                           .when(F.col(column).startswith('T'), 'T')
                           .otherwise(F.col(column))
        ))

#### Criando o dataframe base

#### Função para juntar os groupbys em um dataframe.
#### Uso de union para juntar todos eles

def creating_dataframe(groupedBy, mainDf, column):
    try:
        m_values    = groupedBy.where(F.col(column) == 'M').collect()[0]['count']
    except:
        m_values    = 0
    try:
        f_values    = groupedBy.where(F.col(column) == 'F').collect()[0]['count']
    except:
        f_values    = 0
    try:
        i_values    = groupedBy.where(F.col(column) == 'I').collect()[0]['count']
    except:
        i_values    = 0
    try:
        s_values    = groupedBy.where(F.col(column) == 'S').collect()[0]['count']
    except:
        s_values    = 0
    try:
        t_values    = groupedBy.where(F.col(column) == 'T').collect()[0]['count']
    except:
        t_values    = 0
    try:
        null_values = groupedBy.where(F.col(column).isNull()).collect()[0]['count']
    except:
        null_values = 0    
        
    schema = StructType([
    StructField("qa_",  StringType(),  True),
    StructField("M",    IntegerType(), True),
    StructField("F",    IntegerType(), True),
    StructField("I",    IntegerType(), True),
    StructField("S",    IntegerType(), True),
    StructField("T",    IntegerType(), True),
    StructField("null", IntegerType(), True)
    ])
    
    data          = [(column, m_values, f_values, i_values, s_values, t_values, null_values)]
    df_aux        = spark.getOrCreate().createDataFrame(data = data, schema = schema)
    
    mainDf = mainDf.union(df_aux)
    return mainDf

#columns = ['qa_','M', 'F', 'I', 'S', 'T', 'null']
schema = StructType([
    StructField("qa_",  StringType(),    True),
    StructField("M",    IntegerType(),   True),
    StructField("F",    IntegerType(),   True),
    StructField("I",    IntegerType(),   True),
    StructField("S",    IntegerType(),   True),
    StructField("T",    IntegerType(),   True),
    StructField("null", IntegerType(),   True)
])

dummy  = []
mainDf = spark.getOrCreate().createDataFrame(data = dummy, schema = schema)
#mainDf.show()

contador = 0
for column in df.schema.names:
    if 'qa_' in column:
        groupedBy = df.groupBy(column).count()
        mainDf    = creating_dataframe(groupedBy, mainDf, column)

mainDf.write.mode('overwrite').parquet('../data/datasets_relatorios/qualidade.parquet')

qualidade_data  = '../data/datasets_relatorios/qualidade.parquet'


qualidade = (spark.getOrCreate().read
                  .format("parquet")
                  .option("header", "true")
                  .load(qualidade_data))

qualidade.show(5)

### Checar flights_qa_tailnum, qa_dep_arr, planes_qa_tailnum 

### qa_manufacturer -> classificações C não entraram no dataset
### qa_enginge          -> classificações C não entraram no dataset


qualidade.createOrReplaceTempView('qualidade')

#qualidade.withColumn('check_quantidade',(
#                    qualidade.M + qualidade.F + qualidade.I + qualidade.S + qualidade.T + qualidade.null 
# )).show(100)

#qualidade.show(100)

### Pergunta 3

qualidade.groupby('qa_').agg({'M': 'max'}).sort('max(M)', ascending = False).show(5)

spark.getOrCreate().sql('''
            SELECT *
            FROM qualidade
            ORDER BY M DESC 
            LIMIT 5
''').show()

### Pergunta 4

qualidade.groupby('qa_').agg({'F': 'max'}).sort('max(F)', ascending = False).show(5)

spark.getOrCreate().sql('''
    SELECT * 
    FROM qualidade
    ORDER BY F desc
    LIMIT 5
''').show()

### Pergunta 5

qualidade.groupby('qa_').agg({'I': 'max'}).sort('max(I)', ascending = False).show(5)

spark.getOrCreate().sql('''
        SELECT *
        FROM qualidade
        ORDER BY I desc
        LIMIT 5
''').show()

### Perguntas para Negócio

airports_proc_path_data = '../data/datasets_transformados/airports.parquet'
planes_proc_path_data   = '../data/datasets_transformados/planes.parquet'
flights_proc_path_data  = '../data/datasets_transformados/flights.parquet'


airports_proc = (spark.getOrCreate().read
                  .format("parquet")
                  .option("header", "true")
                  .load(airports_proc_path_data))

planes_proc   = (spark.getOrCreate().read
                  .format("parquet")
                  .option("header", "true")
                  .load(planes_proc_path_data))

flights_proc  = (spark.getOrCreate().read
                  .format("parquet")
                  .option("header", "true")
                  .load(flights_proc_path_data))



airports_origin_proc = airports_proc
airports_dest_proc   = airports_proc


for column_name in flights_proc.schema.names: 
    flights_proc = (flights_proc.withColumnRenamed(column_name, f"flights_{column_name}"))

for column_name in planes_proc.schema.names: 
    planes_proc = (planes_proc.withColumnRenamed(column_name, f"planes_{column_name}"))    

for column_name in airports_proc.schema.names: 
    airports_origin_proc = (airports_origin_proc.withColumnRenamed(column_name, f"origin_{column_name}"))
    airports_dest_proc   = (airports_dest_proc.withColumnRenamed(column_name, f"dest_{column_name}"))

df2 = flights_proc.join(airports_origin_proc, flights_proc.flights_origin == airports_origin_proc.origin_faa, 'left')
df2 = df2.join(airports_dest_proc, df2.flights_dest == airports_dest_proc.dest_faa, 'left')
df2 = df2.join(planes_proc, df2.flights_tailnum == planes_proc.planes_tailnum, 'left' )

df2.write.mode('overwrite').parquet('../data/datasets_relatorios/transformado1.parquet')


df2 = (spark.getOrCreate().read
                  .format("parquet")
                  .option("header", "true")
                  .load('../data/datasets_relatorios/transformado1.parquet'))

df2.createOrReplaceTempView('df2')

#df2.schema.names

### Pergunta 2

df2.groupBy('dest_region', 'origin_region').agg({'dest_name': 'count',
                                                 'origin_name': 'count'
                                                }).show()

df2.groupBy('dest_region').agg(F.countDistinct('dest_region', 'dest_name').alias('count')).show()
df2.groupBy('origin_region').agg(F.countDistinct('origin_region', 'origin_name').alias('count')).show()

print('******SQL******\n')

spark.getOrCreate().sql('''
            SELECT dest_region,
            COUNT (DISTINCT dest_name) as count
            FROM df2
            GROUP BY dest_region
''').show()

spark.getOrCreate().sql('''
            SELECT origin_region,
            COUNT (DISTINCT origin_name) as count
            FROM df2
            GROUP BY origin_region
''').show()

### Pergunta 3

df2 = df2.withColumn('aux_alt',(
                F.when((df2.origin_alt - df2.dest_alt > 0), df2.origin_alt - df2.dest_alt)
                .otherwise(None)
))

#df2.groupBy('dest_region', 'origin_region').agg({'aux_alt': 'max'}).sort('max(aux_alt)', ascending =False).show(5)
df2.select('flights_origin','flights_dest','aux_alt').distinct().sort('aux_alt', ascending = False).show(5)

print('******SQL******\n')

spark.getOrCreate().sql('''
            SELECT 
                distinct 
                flights_origin,
                flights_dest,
                    CASE 
                        WHEN
                            origin_alt - dest_alt > 0 THEN origin_alt - dest_alt
                            ELSE 0
                        END AS diff_alt
            from df2
            ORDER BY diff_alt desc
            limit 5
''').show()

### Pergunta 4

#df2.select(F.mean('flights_dep_delay'), F.when(df2.flights_dep_delay > 0, 1)).show()
#df2.select(F.mean('flights_arr_delay')).show()

df2.where(df2.flights_dep_delay > 0).agg(F.ceil(F.mean('flights_dep_delay')).alias('atraso médio')).show()
df2.where(df2.flights_arr_delay > 0).agg(F.ceil(F.mean('flights_arr_delay')).alias('atraso médio')).show()

print('******SQL******\n')

spark.getOrCreate().sql('''
            SELECT CEIL(AVG(flights_dep_delay))
            FROM df2
                WHERE flights_dep_delay >= 0      
''').show()

spark.getOrCreate().sql('''
            SELECT CEIL(AVG(flights_arr_delay))
            FROM df2
                WHERE flights_arr_delay >= 0      
''').show()

### Pergunta 5

df2.where(df2.flights_dep_delay > 0).groupBy('dest_region', 'origin_region').agg(F.round(F.mean("flights_dep_delay"),2).alias('avg_flights_dep_delay')).show()
df2.where(df2.flights_arr_delay > 0).groupBy('dest_region', 'origin_region').agg(F.round(F.mean("flights_arr_delay"),2).alias('avg_flights_arr_delay')).show()

print('******SQL******\n')

spark.getOrCreate().sql('''
            SELECT distinct dest_region, 
                            origin_region,
                            CEIL(avg(flights_dep_delay)) as avg_flights_dep_delay
            FROM df2 
            WHERE flights_dep_delay >=0
            GROUP BY dest_region, origin_region
''').show()

spark.getOrCreate().sql('''
            SELECT distinct dest_region, 
                            origin_region,
                            CEIL(avg(flights_arr_delay)) as avg_flights_arr_delay
            FROM df2 
            WHERE flights_arr_delay >=0
            GROUP BY dest_region, origin_region
''').show()

### Pergunta 6

df2.where(df2.flights_dep_delay > 0).groupBy(F.year('flights_dep_datetime')).agg({"flights_dep_delay": 'sum'}).show()
df2.where(df2.flights_arr_delay > 0).groupBy(F.year('flights_dep_datetime')).agg({"flights_arr_delay": 'sum'}).show()

print('******SQL******\n')

spark.getOrCreate().sql('''
                SELECT  YEAR(flights_dep_datetime),
                        SUM(flights_dep_delay) as sum_flights_dep_delay
                FROM df2
                where flights_dep_delay >= 0 
                GROUP BY YEAR(flights_dep_datetime)

''').show()

spark.getOrCreate().sql('''
                SELECT  YEAR(flights_dep_datetime),
                        SUM(flights_arr_delay) as sum_flights_arr_delay
                FROM df2
                where flights_arr_delay >= 0 
                GROUP BY YEAR(flights_dep_datetime)

''').show()
#df2.select('flights_dep_datetime').show(1)

### Pergunta 7

df2.where(df2.flights_dep_delay >= 0).groupBy(F.year('flights_dep_datetime'), 'dest_region',).agg({"flights_dep_delay": 'sum'}).show()
df2.where(df2.flights_arr_delay >= 0).groupBy(F.year('flights_dep_datetime'), 'dest_region').agg({"flights_arr_delay": 'sum'}).show()

df2.where(df2.flights_dep_delay >= 0).groupBy(F.year('flights_dep_datetime'), 'origin_region',).agg({"flights_dep_delay": 'sum'}).show()
df2.where(df2.flights_arr_delay >= 0).groupBy(F.year('flights_dep_datetime'), 'origin_region').agg({"flights_arr_delay": 'sum'}).show()

print('******SQL******\n')

spark.getOrCreate().sql('''
            SELECT YEAR(flights_dep_datetime) as year_datetime,
                   dest_region,
                   SUM(flights_dep_delay) as sum_flights_dep_delay
            FROM df2
            WHERE flights_dep_delay >= 0
            GROUP BY YEAR(flights_dep_datetime), dest_region        
''').show()

### Pergunta 8

df2.select(F.ceil(F.mean('flights_air_time'))).show()

print('******SQL******\n')

spark.getOrCreate().sql('''
        SELECT ceil(avg(flights_air_time)) as avg_flights_air_time
        FROM df2
''').show()

### Pergunta 9

df2.groupBy('dest_region',).agg(F.round(F.mean("flights_air_time"),2).alias('avg_dest_flights_air_time')).show()

df2.groupBy('origin_region',).agg(F.round(F.mean("flights_air_time"),2).alias('avg_origin_flights_dep_delay')).show()

print('******SQL******\n')

spark.getOrCreate().sql('''
        SELECT dest_region,
               CEIL(avg(flights_air_time)) as avg_flights_air_time
        FROM df2
        GROUP BY dest_region
''').show()

### Pergunta 10

df2 = df2.withColumn('origin_dest', (
            F.concat_ws('-',df2.flights_origin, df2.flights_dest)
))

df2.groupBy('flights_origin', 'flights_dest').agg(F.ceil(F.mean('flights_air_time'))).show(5)

print('******SQL******\n')

spark.getOrCreate().sql('''
            SELECT flights_origin,
                   flights_dest,
                   ceil(avg(flights_air_time)) as avg_flights_air_time
            FROM df2
            GROUP BY flights_origin, flights_dest
            LIMIT 5
''').show()

### Pergunta 11

df2.groupBy(F.year('flights_dep_datetime')).agg({'flights_air_time': 'sum'}).show()

### Pergunta 12

df2.groupBy('dest_region').agg({'flights_air_time': 'sum'}).show()
df2.groupBy('origin_region').agg({'flights_air_time': 'sum'}).show()

### Pergunta 13

df2.select(F.ceil(F.mean('flights_distance'))).show()

### Pergunta 14

df2.groupBy('origin_region').agg({'flights_distance': 'avg'}).show()
df2.groupBy('dest_region').agg({'flights_distance': 'avg'}).show()

### Pergunta 15

#### Para negócios faz mais sentido ver a média das diferenças ou desvio padrão. Se esses valores não se aproximarem a 0 tem algo de errado.


df2.groupBy('flights_origin', 'flights_dest').agg(F.ceil(F.mean('flights_distance'))).show(5, truncate = False)

### Pergunta 16

df2.groupBy(F.year('flights_dep_datetime')).agg({'flights_distance': 'sum'}).show(5, truncate = False)

### Pergunta 17

df2.groupBy(F.year('flights_dep_datetime'), 'origin_region').agg({'flights_distance': 'sum'}).show(5, truncate = False)
df2.groupBy(F.year('flights_dep_datetime'), 'dest_region').agg({'flights_distance': 'sum'}).show(5, truncate = False)

### Pergunta 18

df2.groupBy('flights_origin', 'flights_dest', 'planes_seats').agg({'planes_seats': 'avg'}).show(5)

### Pergunta 19

df2.groupBy(F.year('flights_dep_datetime')).agg({'planes_seats': 'sum'}).show()

### Pergunta 20

df2.groupBy('flights_dest').agg({'flights_dest': 'count'}).sort('count(flights_dest)', ascending = False).show(5)

### Pergunta 21

df2.groupBy('flights_dest').agg({'planes_seats': 'sum'}).sort('sum(planes_seats)', ascending = False).show(5)

### Pergunta 22

df2.groupBy('flights_origin', 'flights_dest').agg({'flights_distance': 'max'}).filter(df2.flights_origin == 'PDX').sort('max(flights_distance)', ascending = False).show(5)

### Pergunta 23

df2.groupBy('flights_dest', F.month('flights_dep_datetime').alias('month')).count().sort('count', ascending = False).show(5)

### Pergunta 24

df2.groupBy('planes_model').count().sort('count', ascending = False).show(5)

### Pergunta 25

df2.where(df2.planes_model.isNotNull()).groupby('planes_model', 'flights_dest').count().sort('count', ascending = False).show(5)

### Pergunta 26

df2.where(df2.planes_engines.isNotNull()).groupBy('flights_haul_duration').agg(F.ceil(F.mean(df2.planes_engines)).alias('avg_engines')).sort('avg_engines').show()

### Pergunta 27

df2.groupBy('flights_dep_season').agg({'flights_dep_season': 'count'}).show()

### Pergunta 28

df2.groupBy('flights_dep_season', 'flights_dest').count().sort('count', ascending = False).show()

### Pergunta 29

df2.where(df2.flights_dep_delay_category != 'ANTECIPATED').groupBy('flights_dep_delay_category').count().show()

### Pergunta 30

df2.where(df2.flights_dep_delay_category != 'ANTECIPATED').groupBy('flights_dep_delay_category', 'flights_dest', 'flights_origin').count().sort('count', ascending = False).show(5)
